+++ testing/test_runner.py
@@ -568,6 +568,32 @@ def test_makereport_getsource(testdir):
     result.stdout.fnmatch_lines(['*else: assert False*'])
 
 
+def test_makereport_getsource_dynamic_code(testdir, monkeypatch):
+    """Test that exception in dynamically generated code doesn't break getting the source line."""
+    import inspect
+    original_findsource = inspect.findsource
+    def findsource(obj, *args, **kwargs):
+        # Can be triggered by dynamically created functions
+        if obj.__name__ == 'foo':
+            raise IndexError()
+        return original_findsource(obj, *args, **kwargs)
+    monkeypatch.setattr(inspect, 'findsource', findsource)
+
+    testdir.makepyfile("""
+        import pytest
+
+        @pytest.fixture
+        def foo(missing):
+            pass
+
+        def test_fix(foo):
+            assert False
+    """)
+    result = testdir.runpytest('-vv')
+    assert 'INTERNALERROR' not in result.stdout.str()
+    result.stdout.fnmatch_lines(["*test_fix*", "*fixture*'missing'*not found*"])
+
+
 def test_store_except_info_on_eror():
     """ Test that upon test failure, the exception info is stored on
     sys.last_traceback and friends.